requirements.txt
torch
torchvision
ninja
yacs
cython
matplotlib
tqdm
opencv-python
scikit-image
tensorboard
graphviz
